refactor: introduce post orchestrator boundary
